use restrict/static
I see these as explaining the behavior of functions through semantics
rather than documentation. `src` and `dst` parameters can't point to the
same object if `restrict` is used. Pointers can't be null if `static 1`
is used in a function parameter array declarator.

These are rarely used, but I expect other C programmers to read the
standard if it causes confusion.